Rename servermcp package to mcpserver
